LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/errorhandling-expected.txt

CONSOLE MESSAGE: TypeError: 'text/plain' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'text/plain' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: SyntaxError: No identifiers allowed directly after numeric literal
CONSOLE MESSAGE: SyntaxError: No identifiers allowed directly after numeric literal
CONSOLE MESSAGE: SyntaxError: No identifiers allowed directly after numeric literal
CONSOLE MESSAGE: SyntaxError: No identifiers allowed directly after numeric literal
html-script-module-errorHandling


FAIL IFrame test: 'iframe_parseError_Root' assert_equals: Unexpected _errorReported value expected "error" but got ""
PASS IFrame test: 'iframe_parseError_Dependent'
PASS IFrame test: 'iframe_parseError_DependentMultiple'
PASS External root module with non-script mimetype
PASS Module with imported non-script mimetype
